refactor(table-service): #15009 outsource it
Additionally move the filter method there
